Footer updates and JS error handling
Made footer fixed to the bottom of the window, with JavaScript to handle overlap between the content box and footer that creates a background and blurs content behind the footer. Removed the spacer from pages that was doing nothing, CSS changes - Removed min-height from body, adjusted margins for dark theme switch. Fixed javascript errors of searching for elements on pages they aren't on.